add db and auth

+++ src/api/notes/handler.js
@@ -16,8 +16,11 @@ class NotesHandler {
     try {
       this.validator.validateNotePayload(request.payload);
       const { title = 'untitled', body, tags } = request.payload;
+      const { id: credentialId } = request.auth.credentials;
 
-      const noteId = await this.service.addNote({ title, body, tags });
+      const noteId = await this.service.addNote({
+        title, body, tags, owner: credentialId,
+      });
 
       const response = h.response({
         status: 'success',
@@ -47,8 +50,9 @@ class NotesHandler {
     }
   }
 
-  async getNotesHandler() {
-    const notes = await this.service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this.service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -60,6 +64,9 @@ class NotesHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this.service.verifyNoteAccess(id, credentialId);
       const note = await this.service.getNoteById(id);
       return {
         status: 'success',
@@ -92,6 +99,9 @@ class NotesHandler {
     try {
       this.validator.validateNotePayload(request.payload);
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this.service.verifyNoteAccess(id, credentialId);
 
       await this.service.editNoteById(id, request.payload);
 
@@ -123,6 +133,9 @@ class NotesHandler {
   async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this.service.verifyNoteOwner(id, credentialId);
       await this.service.deleteNoteById(id);
       return {
         status: 'success',
